chg: use only floats in bvh update

+++ mujoco_contact_surface_sensors/include/mujoco_contact_surface_sensors/flat_tactile_sensor.h
@@ -159,7 +159,7 @@ protected:
 	void mt_update(const mjModel *m, mjData *d, const std::vector<GeomCollisionPtr> &geomCollisions);
 	void bvh_update(const mjModel *m, mjData *d, const std::vector<GeomCollisionPtr> &geomCollisions);
 
-	void render_tiles(Eigen::ArrayXXd pressure, mjtNum rot[9], mjtNum origin[3]);
+	void render_tiles(Eigen::ArrayXXf pressure, mjtNum rot[9], mjtNum origin[3]);
 
 private:
 	bool use_parallel       = true;
